Merge branch '652227-directories' and several smaller branches from Irakli.

This changes the linker-side module search behavior in several ways, most
importantly improving things like allowing explicit require(addon-kit/panel)
avoid doing a search, and allowing require(foo) to find the "main" entry
point of a package named foo. It also changes the JS-side runtime search
behavior (used mostly by unit test code when it calls require() ) to do
mostly the same thing. Documentation on what it now does are in
static-files/md/dev-guide/addon-development/module-search.md .

This should resolve the following bugs:

 Bug 652227 : change ".directories" handling to tolerate NPM modules with
              JS in root directory
 Bug 654983 : explicit require with PKG/MOD/PATH does not work
 Bug 658198 : make all require args relative or PKG/MOD style for modules
              from other packages
 Bug 556582 : Module path search order is undefined/undocumented

and resolve or at least improve these other bugs:

 Bug 606597 : conform to CommonJS Packages and Modules specifications
 Bug 614707 : let Jetpack use CommonJS and NPM modules
 Bug 552841 : confirm to CommonJS Package 1.1 standard

+++ python-lib/cuddlefish/tests/test_xpi.py
@@ -27,7 +27,7 @@ class Bug588119Tests(unittest.TestCase):
     def tearDown(self):
         if self.xpi:
             self.xpi.close()
-        if self.xpiname:
+        if self.xpiname and os.path.exists(self.xpiname):
             os.remove(self.xpiname)
 
     def testPackageWithImplicitIcon(self):
